backend/requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.28.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# AWS SDK
boto3>=1.28.0
requests>=2.31.0
httpx>=0.24.0

# Security and Encryption
cryptography>=41.0.0

# Pydantic
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
email-validator>=2.0.0
